backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
celery[redis]
pydantic
python-dotenv
pymysql
python-jose[cryptography]
passlib[bcrypt]
Pillow
opencv-python-headless
numpy
openpyxl
python-multipart
paddleocr>=2.6.0
pydantic-settings
